requirements.txt
fastapi==0.116.1
uvicorn[standard]
python-dotenv

python-multipart
httpx

# extraction fichiers
pdfplumber==0.11.0
python-docx==1.1.0
pytesseract
pdf2image
Pillow


python-jose[cryptography]
passlib
bcrypt==4.0.1


chromadb

sentence-transformers
torch


protobuf
sentencepiece
tiktoken

weasyprint
jinja2
openpyxl

rank-bm25
